Activate querying of the Tiro voices API when VoiceManager is started.
So far no visible feedback to GUI, but nicely working logs in logcat.
When no network is available, the error callback is not called, though.
Therefore, we need to think about enable/disable flag in the Voice model
for network voices, that is only set to enabled, if the API returns a
successful query for the voices.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>