Implement playing of Tiro network voices
The voices can now be played from Tiro API.

We are querying the currently chosen voice from the Voice Model and query
current network status. If the network status is good, we enable the Speak button
and change the Status icon to cloud_checked icon.

When pressing the play button, a request is generated with the currently
chosen VoiceId, text of the edit textbox, mp3 format, 22.05 kHz sample
frequency and is then sent to Tiro TTS API. The returned audio is played
via the internal Android Media player.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>