Revamp FliteTTS Service to be soley Model-based.
This makes it possible, to use the Tiro network voices for Icelandic locales
for Android TTS.

Cut dependency to Flite Voices. It's only relevant now, if a voice is
inside the AppRepository and therefore, inside the DB.

We can reintroduce support of FLite Voices, if these are managed by
the AppRepository as well.

Now we only have to replace the Flite voice dependency in
CheckVoiceData.java, to also be able to choose the appropriate voice out of
our model

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>