Implement all TTS service methods for network voices
Androids approach for multiple voices per language/country/variant is not
really easy to grasp. The implementation of this commit uses the voice name
as locale variant. This variant is shown in the voice chooser settings
screen. If the user chooses a voice, onLoadVoice() is executed.

When onLoadVoice() is called, we cache the selected voice in the member
variable AppRepository::mLoadedVoice, but also write the voice id into
the AppData db entry. The next time onGetDefaultVoiceNameFor() is called,
we either return what is inside AppRepository::mLoadedVoice or return the
voice name by using the voice id returned by mAppDataDao.getCurrentVoiceId().

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>