Activate icelandic Android TTS service via network voices (#17)
This MR activates the Android TTS service for Icelandic. The activated voices are soley network based for the moment via the Tiro Speech API. We also do activate normalization with the new NormalizationManager class before the text leaves the phone.
At the same time, we disable FliteTTS, because this currently only includes non-Icelandic voices and were only for demo- and functionality proof for the FLite native part. We will active FLite again, once local Icelandic voices are supported.

Androids approach for multiple voices per language/country/variant is not really easy to grasp. This implementation uses the voice name as locale variant. This variant name is shown in the voice chooser settings screen.

If the user chooses a voice, TTSService::onLoadVoice() is executed. When onLoadVoice() is called, we cache the selected voice in the member variable AppRepository::mLoadedVoice, but also write the voice id into the AppData db entry.

The next time onGetDefaultVoiceNameFor() is called, we either return what's inside AppRepository::mLoadedVoice or return the voice name by using the voice id returned by member mAppDataDao.getCurrentVoiceId().

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>