Introduce class TiroServiceGgenerator: use static initialization for http client
All members of the class are statically initialized and therefore cached
for the lifetime of the application. This way, we use the same http client
for all requests to Tiro API.
Furthermore explicitely use the order of protocols http_2, http_1_1. okhttp
will use automatically http2, if server supports it.
Enable also the logging interceptor and prepare TiroServiceGgenerator for
token based authentication.

Adapt SpeakController/VoiceController to use the new class.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>