Catch exception when trying to update the database.
When analyzing the Google play installed application, empty voice entries
were generated by the Retrofit http client in combination with GSON entries.
These led to a crash, because the DB layer complained about missed non-
null constrains. Catch the exception and log it.

Additionally always request new voice list, if the VoiceManager activity is
started. With that, we can trigger a network request instead of waiting until
the voice update timer expired, which is currently 30 minutes.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>