SpeakController/VoiceController: fix bug reported in Firebase
When calling stop(), ongoing calls were cancelled and the observers and
Retrofit calls were nulled. But the callbacks onFailure() method is
nevertheless executed from Retrofit as a consequence of calling
call.cancel(). Therefore, we got a NPE when trying to access the observers
error() method inside onFailure().

The solution implemented in this commit doesn't null the calls anymore and
simplifies the onFailure() methods to their bare minimum.
In case the call is cancelled, no observer error() method is called.

This behaviour can be tested by using the Screen-Reader and rapidly
skipping the currently played Screenpart to the next section.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>