VoiceInfo: show cancellable dialog, when TTS service/network connection isn't available.
In case no internet or TTS service backend is available, instead of silently
failing, a dialog is shown, that there is either no internet connection
at all, or the TTS service backend is not reachable. In case of the former,
the user is directed to the Wifi settings activity in case of the latter,
the user is hinted to try later.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>