app/build.gradle: save last modification time in assets/lastModified.txt
In order to determine, when to copy over assets to the application's data
directory, we can utilize this file. If any assets are listed as newer
than the modification timestamp on the Android device, we should copy it,
otherwise, we just leave the file as is.
This procedure has still to be implemented, but it's good to bundle this
file already with the new release.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>