frontend/Pronunciation.java: lazy load G2P otherwise test cases fail to run
We have a dependency problem for G2P in test cases. G2P needs native
methods, which we don't have for the native platform running the unit tests.
As so far we don't have unit tests involving G2P or other native dependencies,
we simply lazy load G2P unless we really need it (i.e. the first time, a
on-device voice is needed)

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>