2nd part of the utterance cache handling.
Add caching to all voice requests. Before network and on-device voices are
synthesized, the raw text is looked up in the utterance cache. Normalization
and G2P is cached indpendently from audio data. Frontend pipeline results
and voice speech audio data is attached to an utterance cache item.

If audio has already been synthesized and the cache item is not expired,
the audio data is loaded from the cache directory and played directly.

For all "hot" cache items the delay between onSynthesizeText() and
playback of speech audio  is reduced to ~0.1 seconds. This makes it
possible to quickly skip audio playback.

More details:

- add x-request-id header to network voice requests to identify the
  associated cache item for attaching audio
- if utterance should be stopped, also honor stopping inside registered
  observers
- centralize Normalization manager and keep only one instance around

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>