Rework speech audio playback and service callback feeding.
Callbacks are now never feeded from within anywhere else than the TTS
onSynthesizeText() context. Before, async threads were used. At the same
time, remove the synchronized attribute for the methods onSynthesizeText()
and onStop(), because otherwise it's not possible to receive the stopping
request of an ongoing utterance.

The processing of audio is still asynchronous, but results as well as
stop requests are feeded into a queue and the onSynthesizeText() method
now waits either on the text audio data or on stopping requests or
error messages. For these, elements of the new class TTSProcessingResult
are enqueued.

This addresses the following issues:

- synchronization problems encountered when fast skipping over cached audio
- unstoppable text audio playing

Add TTSRequest class for distinguishing the same utterances but different requests

This fixes a race condition, where the same utterance could have been
requested after the current has been stopped. In that case, there was no
possibility to distinguish the new from the cancelled one. Therefore,
a TTSRequest object identifies the utterance id and a unique request id.

The combination of both has also to be sent to the Network API as
x-request-id header, so that we can assign the correct response
and similarly distinguish the same utterances.

Furthermore simplified speech mark handling has been implemented.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>