Generalize Network voice handling
Use api.grammatek.com/tts for the network voice request and adapt all
classes for the more general term 'network'. It's easy to switch the
network endpoint URL.

Don't bail out on non-existence of the x-request-id response from server.

Add database migration step to drop all entries with type 'tiro'. As the
voice list is updated from the new network endpoint, it will be newly
populated with the 'network' type.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>