Add configuration settings rm_cache_item_after_playing, rm_cache_item_for_fast_voices
These are foremost helpers for debugging so that we don't need to manually
delete the application data, if we try out new versions of frontend and/or
voice audio.
But of course there is no need for caching in case the voice generation
on the phone is fast enough, so we could also activate rm_cache_item_for_fast_voices
in combination with a good RTF setting even for production.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>