Revamp voice version handling and cache cleanup
This is a bug fix for deleting cached utterances in case of fast voices.
Because of the calculation of the RTF (realtime factor) even for the case
of reading from cached utterances, any cache item was deleted after being
played, if the setting 'rm_cache_item_for_fast_voices' was set to true,
which is the default.
Furthermore, the missing version info for the cache item caused cache handling
bugs that took the version info into account.

- don't delete cache items in case it's already inside the cache
- add sanity check for RTF-calculations > 500, which are a hint on
  either cancellations or problems and don't delete the cache item in that
  case
- if the time measured between enquing a processing request and the
  processing response is 0, don't calculate RTF as infinity
- always add the version of the voice for any cache item
- add dummy version info v0 for network voices
- migrate any existing network voice version in the db to v0
- assert on missing version info
- add new db migrations 6->7 for the voices table
- fix missing migration tests

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>