Fixes the state "Server Voice description not available !"
If the initial request for fetching the voice description was not successful,
the message "Server Voice description not available !" would only disappear
either after a restart of the app and a hopefully then successful fetching
of the voice info from the server or after the interval of 60 minutes for
the next server voice info update had been reached. This commit fixes the
problem that downloading the voice "Álfur hratt" doesn't work in the
meanwhile.

Add a new trigger method "triggerServerVoiceUpdate()" for downloading the
server voice info to the AppRepository class. This method initiates async.
downlad of the server voice info in the usual way.
The call is done locally inside readVoiceDescription(false), i.e for the
case that includeServer parameter is set to false.

Basically this trigger is an indirect recursion of
readVoiceDescription(false) to readVoiceDescription(true) via an async.
scheduler call to circumvent blocking the UI thread.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>