Correctly handle EMail intent
Before launching the EMail SENDTO intent, we need to check if there is an
activity listening to it. Since Android 11, for that to be permitted, we
need to put an appropriate <queries> ... </queries> permission into the
manifest.

This fixes the seldomly found in real-world scenarios, but always existing
crash in Pre-Launch reports for Android 8 devices:

java.lang.NullPointerException: Attempt to invoke virtual method
 'android.accounts.Account com.android.mail.providers.Account.b()'
  on a null object reference

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>