Don't do any connectivity check and only accept ONNX voices in assets
Initiate searching for voices in assets immediately at the start of
the App, but don't do any connection check to Grammatek TTS API and also
don't get the release info from Github.

Also fix the error in mapDeviceVoiceToDbVoice() for comparing the asset
voice name to the one inside the DB, so that the voice can be loaded
immediately.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>