TTSEngineOnnx: add sentence-wise reading of Strings
In case the given text is made up of multiple sentences, split it according
to the characters `.!?;`. Otherwise, the current voice reads simply over
these symbols without pause, when these are not found at the end of the
text.

Also add method Audiomanager#generatePcmSilence() that takes a duration
and returns an appropriate byte[] buffer of 0 bytes.

Reuse the new method also in

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>