Database Migration 7 => 8 fix empty voice_table and current_voice_id
In case where the voice_table is empty, an update of the column
current_voice_id wouldn't be possible in combination with the NON NULL
constraint. Take this case into account and set the value then to -1.

User code has to handle that scenario that basically means: no voice
is currently selected. Currently there is no user code dependend on this.

Add an appropriate migration test.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>