Use the voice settings for the pitch.
In contrast to the speech rate, there is no setting in e.g. Talkback to
change the pitch on-the-fly.
Somehow and only for some phones, we get nevertheless strange pitch
variations per request and unfortunately cannot trust it. Therefore,
restore the previous pitch setting behaviour and use the TTS settings menu
pitch value of the phone as has been done before for Símarómur versions
2.0.2 and earlier.

See also issue #24

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>